pyproject.toml
[tool.poetry]
name = "cloud-instance"
version = "0.1.5"
description = ""
authors = ["Fabio Ghirardello"]
license = "GPLv3+"
readme = "README.md"
homepage = "https://github.com/fabiog1901/cloud_instance"
repository = "https://github.com/fabiog1901/cloud_instance"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent"
]

[tool.poetry.scripts]
cloud_instance = 'cloud_instance.cli.cli:app'

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.38.8"
boto = "^2.49.0"
botocore = "^1.38.8"
google-api-core = "^2.24.2"
google-auth = "^2.39.0"
google-cloud-compute = "^1.30.0"
googleapis-common-protos = "^1.70.0"
azure-common = "^1.1.28"
azure-core = "^1.34.0"
azure-identity = "^1.21.0"
azure-mgmt-compute = "^34.1.0"
azure-mgmt-core = "^1.5.0"
azure-mgmt-network = "^28.1.0"
azure-mgmt-resource = "^23.3.0"
typer = "^0.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
